Allow personnel to belong to many squadrons
Instead of a one-many relation where a person can only belong to one
squadron, a many-many relation is now used where a person can belong to
multiple (or zero) squadrons.

This simplifies squadron-editing because it avoids needing to check
uniqueness when adding people to squadrons (i.e. no need to check that
a person is unique across all squadrons).

* New join table PersonnelSquadrons added to schema
* Person class now has a Squadrons list instead of a single Squadron
* Squadron personnel are now populated in a similar manner to KeyBunch's
  AuthorizedPersonnel, by iterating over a LEFT OUTER JOIN result and
  adding to the collection
* Adding personnel to a Squadron now simply involves adding to
  Squadron.Personnel, instead of calling a special method
* Grabbing a new list of squadrons from the database now requires
  calling Squadron.ById() to fully populate the personnel list